Switched to f64 & some basic refactoring.

+++ src/structs/types.rs
@@ -2,7 +2,7 @@ use nalgebra::Vector3;
 use nalgebra::Vector2;
 
 pub type Color = (u8, u8, u8);
-pub type Point = Vector2<i32>;
+pub type Point = Vector2<f64>;
 pub type Line = (Point, Point);
-pub type Vec3f = Vector3<f32>;
+pub type Vec3f = Vector3<f64>;
 pub type Vec3i = Vector3<i32>;
